added a experience progress bar to the user page to show amount of exp until the next level

+++ Group14-Project/src/main/java/com/example/group14project/controller/UserController.java
@@ -49,7 +49,7 @@ public class UserController {
 
         int level = user.getCurrentLevel();
         model.addAttribute("level", level);
-        float exp = user.getTotalExp();
+        int exp = user.getTotalExp();
         model.addAttribute("exp", exp);
 
         return "user";
